defined kfd/ops.py
